shift to dashboard view
