+++ src/features/user/TreeMapper/MySpecies/MySpecies.module.scss
@@ -1,4 +1,4 @@
-@import '../../../theme/theme';
+@import '../../../../theme/theme';
 
 .mySpeciesContainer {
   display: grid;
