+++ src/features/user/TreeMapper/MySpecies/index.tsx
@@ -1,0 +1,19 @@
+import React, { ReactElement } from 'react';
+import DashboardView from '../../../common/Layout/DashboardView';
+import CenteredContainer from '../../../common/Layout/CenteredContainer';
+import MySpeciesForm from './MySpeciesForm';
+import { useTranslation } from 'next-i18next';
+
+interface Props {}
+
+export default function MySpecies({}: Props): ReactElement {
+  const { t } = useTranslation('me');
+
+  return (
+    <DashboardView title={t('me:mySpecies')} subtitle={null}>
+      <CenteredContainer>
+        <MySpeciesForm />
+      </CenteredContainer>
+    </DashboardView>
+  );
+}
